pyproject.toml
[tool.poetry]
name = "microgo"
version = "0.1.0"
description = ""
authors = ["highlander-spirou <highlandetspirou2812@gmail.com>"]
readme = "README.md"
packages = [
    { include = "file_access", from="lib" },
    { include = "message_queue", from="lib" }
]

[tool.poetry.dependencies]
python = "^3.10"


[tool.poetry.group.auth-srv.dependencies]
pyjwt = "^2.8.0"
flask = "^3.0.0"
sqlalchemy = "^2.0.23"
mysql-connector-python = "^8.2.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.gateway-srv.dependencies]
flask = "^3.0.0"
pika = "^1.3.2"
python-dotenv = "^1.0.0"
flask-pymongo = "^2.3.0"
requests = "^2.31.0"


[tool.poetry.group.img-srv.dependencies]
pillow = "^10.1.0"



[tool.poetry.group.celery-dispatcher.dependencies]
celery = "^5.3.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
